[svn] delete keywords or categories in metadata edit panel with del-key

+++ ImageMetadataViewer/src/de/elmar_baumann/imv/controller/filesystem/ControllerDeleteFiles.java
@@ -59,7 +59,7 @@ public class ControllerDeleteFiles extends Controller implements ActionListener 
     @Override
     public void actionPerformed(ActionEvent e) {
         Content content = thumbnailsPanel.getContent();
-        if (isControl() && (content.equals(Content.Directory) ||
+        if (isControl() && thumbnailsPanel.hasFocus() && (content.equals(Content.Directory) ||
             content.equals(Content.FavoriteDirectory) ||
             popup.isDeleteFiles(e.getSource()))) {
             deleteSelectedFiles();
